requirements.txt
aiohttp>=2.2.5
alphabet_detector
beautifulsoup4
bleach
coffeehouse
emoji
feedparser
future
geopy
git+https://github.com/starry69/python-markdown2.git
Pillow>=4.2.0
pretty_errors
psutil
pynewtonmath
spongemock
wikipedia
cachetools
CurrencyConverter
googletrans==4.0.0-rc1
humanize
jikanpy
lxml
nltk
psycopg2-binary
pyrate-limiter
python-telegram-bot==13.7
regex
requests
spamwatch
speedtest-cli
sqlalchemy
telethon==1.16.4
ujson
zalgo-text
fontTools
pyrogram
nekos.py
telegraph
tgcrypto
youtube_dl
youtube_search
pytube
youtube_search_python
pymongo
motor
dnspython
redis
gtts
cloudscraper
wget
tswift
lyricsgenius
envparse
cryptg
cryptography
python-arq
aioredis
odmantic
aiohttp[speedups]
pykeyboard
aiofiles
hachoir
carbonnow
bing_image_downloader
search_engine_parser
pytz
html2text
NumPy
opencv-python-headless
uvloop
python-dotenv
gpytranslate
glitch_this
countryinfo
faker
PyDictionary==2.0.1
ffmpeg
python-rapidjson
heroku3
git+https://github.com/python-telegram-bot/ptbcontrib.git
markdown2>=2.4.0
cowpy
selenium
httpx[http2]>=0.18.1,<1.0.0
hiredis
pendulum
img2pdf
nhentai
